View表示の共通処理を作成
