+++ app/javascript/view_helper.js
@@ -1,0 +1,24 @@
+export function displayView(view, parentId, position) {
+  const target = document.getElementById(parentId)
+  target.insertAdjacentHTML(position, view)
+}
+
+export function escapeHTML(strings, ...values) {
+  return strings.reduce((result, str, i) => {
+    const value = values[i - 1];
+    if (typeof value === "string") {
+      return result + escapeSpecialChars(value) + str;
+    } else {
+      return result + String(value) + str;
+    }
+  });
+}
+
+function escapeSpecialChars(str) {
+  return str
+    .replace(/&/g, "&amp;")
+    .replace(/</g, "&lt;")
+    .replace(/>/g, "&gt;")
+    .replace(/"/g, "&quot;")
+    .replace(/'/g, "&#039;");
+}
